Move EntityUploader
